+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/bean/entity/HibernateSenderInfo.java
@@ -17,7 +17,8 @@ public class HibernateSenderInfo implements Bean {
 
     private static final long serialVersionUID = 6717479113778427998L;
 
-    // -----------------------------------------------------------主键-----------------------------------------------------------
+    // region 主键
+
     @Id
     @Column(name = "notify_setting_id", nullable = false)
     private Long notifySettingId;
@@ -26,7 +27,10 @@ public class HibernateSenderInfo implements Bean {
     @Column(name = "topic_id", length = Constraints.LENGTH_ID, nullable = false)
     private String topicId;
 
-    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    // endregion
+
+    // region 主属性字段
+
     @Column(name = "label", length = Constraints.LENGTH_LABEL)
     private String label;
 
@@ -39,7 +43,10 @@ public class HibernateSenderInfo implements Bean {
     @Column(name = "remark", length = Constraints.LENGTH_REMARK)
     private String remark;
 
-    // -----------------------------------------------------------多对一-----------------------------------------------------------
+    // endregion
+
+    // region 多对一
+
     @ManyToOne(targetEntity = HibernateNotifySetting.class)
     @JoinColumns({ //
             @JoinColumn(name = "notify_setting_id", referencedColumnName = "id", insertable = false, updatable = false), //
@@ -52,7 +59,10 @@ public class HibernateSenderInfo implements Bean {
     })
     private HibernateTopic topic;
 
-    // -----------------------------------------------------------审计-----------------------------------------------------------
+    // endregion
+
+    // region 审计
+
     @DatamarkField(handlerName = "senderDatamarkHandler")
     @Column(
             name = "created_datamark",
@@ -68,10 +78,13 @@ public class HibernateSenderInfo implements Bean {
     )
     private String modifiedDatamark;
 
+    // endregion
+
     public HibernateSenderInfo() {
     }
 
-    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
+    // region 映射用属性区
+
     public HibernateSenderInfoKey getKey() {
         return new HibernateSenderInfoKey(notifySettingId, topicId);
     }
@@ -86,7 +99,10 @@ public class HibernateSenderInfo implements Bean {
         }
     }
 
-    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
+    // endregion
+
+    // region 常规属性区
+
     public Long getNotifySettingId() {
         return notifySettingId;
     }
@@ -166,6 +182,8 @@ public class HibernateSenderInfo implements Bean {
     public void setModifiedDatamark(String modifiedDatamark) {
         this.modifiedDatamark = modifiedDatamark;
     }
+
+    // endregion
 
     @Override
     public String toString() {